+++ PathTracer/Utils/Transform.h
@@ -12,16 +12,16 @@
 #include "typeAlias.h"
 #include "Ray.h"
 
-inline Point3f worldToLocal(Matrix4f invM, Point3f pW) {
+inline Point3d worldToLocal(Matrix4f invM, Point3d pW) {
     Point4f homoPW(pW(0), pW(1), pW(2), 1.0);
     Point4f homoPL = invM * homoPW;
-    return Point3f(homoPL(0), homoPL(1), homoPL(2));
+    return Point3d(homoPL(0), homoPL(1), homoPL(2));
 }
 
-inline Point3f localToWorld(Matrix4f M, Point3f pL) {
+inline Point3d localToWorld(Matrix4f M, Point3d pL) {
     Point4f homoPL(pL(0), pL(1), pL(2), 1.0);
     Point4f homoPW = M * homoPL;
-    return Point3f(homoPW(0), homoPW(1), homoPW(2));
+    return Point3d(homoPW(0), homoPW(1), homoPW(2));
 }
 
 class Transform {
@@ -37,11 +37,11 @@ public:
     
     void calcInvM();
     void calcInvMRot();
-    void setTranslation(float* ds);
-    void setScale(float* s);
-    void setRotateX(float rad);
-    void setRotateY(float rad);
-    void setRotateZ(float rad);
+    void setTranslation(double* ds);
+    void setScale(double* s);
+    void setRotateX(double rad);
+    void setRotateY(double rad);
+    void setRotateZ(double rad);
     void setIdentityM();
     
     Matrix4f getM() const;
